LayoutTests/imported/w3c/web-platform-tests/permissions/permissions-cg.https.html

<!doctype html>
<meta charset=utf-8>
<title>Test PermissionStatus's name attribute.</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script>
async function garbageCollect() {
  if (typeof TestUtils !== 'undefined' && TestUtils.gc) {
    await TestUtils.gc();
  } else if (self.gc) {
    await self.gc();
  } else if (self.GCController) {
    // Present in some WebKit development environments
    await GCController.collect();
  } else {
    var gcRec = function (n) {
      if (n < 1)
        return {};
      var temp = {i: "ab" + i + (i / 100000)};
      temp += "foo";
      gcRec(n-1);
    };
    for (var i = 0; i < 1000; i++)
      gcRec(10);
  }
}

promise_test(async () => {
  const { state: initialState } = await navigator.permissions.query({
    name: "geolocation",
  });

  const pass = new Promise(async (resolve) => {
    let status = await navigator.permissions.query({
      name: "geolocation",
    });
    status.addEventListener("change", resolve, { once: true });

    status = null;
    garbageCollect();

    // Will cause change event to be dispatched.
    test_driver.set_permission({ name: "geolocation" }, "granted");
  });

  // Wait for the change event to be dispatched.
  await pass;

  // Reset the permission to its default value.
  await test_driver.set_permission({ name: "geolocation" }, initialState);
}, "status is not garbage collected when it goes out of scope");

</script>
