LayoutTests/imported/w3c/web-platform-tests/permissions/permissions-cg.https-expected.txt


PASS status is not garbage collected when it goes out of scope
